chore: rename reward worker and reorganize code
