+++ examples/embodiment/train_embodied_agent.py
@@ -23,7 +23,7 @@ from rlinf.runners.embodied_runner import EmbodiedRunner
 from rlinf.scheduler import Cluster
 from rlinf.utils.placement import HybridComponentPlacement
 from rlinf.workers.env.env_worker import EnvWorker
-from rlinf.workers.reward.reward_worker import RewardInferenceWorker
+from rlinf.workers.reward.reward_worker import EmbodiedRewardWorker
 from rlinf.workers.rollout.hf.huggingface_worker import MultiStepRolloutWorker
 
 mp.set_start_method("spawn", force=True)
@@ -78,7 +78,7 @@ def main(cfg) -> None:
     if cfg.get("reward", {}).get("use_reward_model", False):
         # Create reward worker group
         reward_placement = component_placement.get_strategy("rollout")
-        reward_group = RewardInferenceWorker.create_group(cfg).launch(
+        reward_group = EmbodiedRewardWorker.create_group(cfg).launch(
             cluster, name=cfg.reward.group_name, placement_strategy=reward_placement
         )
